% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demiter.R
\docType{package}
\name{demiter}
\alias{demiter}
\title{demiter: Iterators for demographic arrays}
\description{
Iterators for demographic arrays, plus
classes and functions for workign with
these iterators. End users would not
normally use this package directly.
}
\details{
The iterators themselves are built from environments.
They break with R's usual semantics in that they
are modified in place by the 'next' functions.
Calling a 'next' function on an interator that
has been exhausted throws an error.

All iterators have a has_next function, which
returns FALSE if the iterator has been exhausted.

Iterators are created from objects with class SpecIter,
by a call to an 'iter_create' function. 

\pkg{demiter} does not use formal classes and
methods for iterators, but instead uses
suffixes on all functions.
}
