% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Iter-generators.R
\name{IterIncrement}
\alias{IterIncrement}
\title{Create an object of class "IterIncrement"}
\usage{
IterIncrement(
  dim_self,
  dim_oth,
  map_dim,
  comp_type_self,
  i_triangle_self = NULL,
  indices_orig_self = NULL,
  indices_dest_self = NULL,
  i_direction_self = NULL
)
}
\arguments{
\item{dim_self}{An integer vector. The dimensions of array \code{self}.}

\item{dim_oth}{An integer vector. The dimensions of array \code{oth}.}

\item{map_dim}{An integer vector mapping the dimensions of \code{self}
on to those of \code{oth}. If a dimension from \code{self} is not
found in \code{oth}, the associated element of \code{map_dim} is \code{0}.}

\item{comp_type_self}{The type of component: \code{"increment"},
\code{"decrement"}, \code{"orig-dest"}, or \code{"pool"}.}

\item{i_triangle_self}{The index of any dimension of array \code{self}
that has dimtype \code{"triangle"}. Equal to \code{NULL} if \code{self}
does not have a triangle dimension.}

\item{indices_orig_self}{An integer vector containing
the indices of any dimensions of \code{self} with dimtype
\code{"origin"}.}

\item{i_direction_self}{The index of any dimension of array \code{self}
that has dimtype \code{"direction"}. Equal to \code{NULL} if \code{self}
does not have a direction dimension.}

\item{indices_orig_dest}{An integer vector containing
the indices of any dimensions of \code{self} with dimtype
\code{"destination"}.}
}
\value{
An object of class \code{IterIncrement}.
}
\description{
Create an object of class \code{IterCollpse} that
contains the information needed to create an iterator
that maps cells in array \code{self} to cells
in arrays of increments or decrements.
}
\examples{
x <- IterCohort(dim = c(4, 2, 3),
                i_time = 3,
                i_age = 1,
                i_triangle = 2
                stop_at_oldest = FALSE,
                offset = 0)
}
\seealso{
To create an increment iterator from an object of class
\code{IterIncrement}, use function \code{\link{create_iter_increment}}.
}
