% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpecIter-generators.R
\name{SpecIterAccount}
\alias{SpecIterAccount}
\title{Create an object of class "SpecIterAccount"}
\usage{
SpecIterAccount(dim, i_time, i_age = NULL)
}
\arguments{
\item{dim}{An integer vector. The dimensions of the array.}

\item{i_time}{The index for the time dimension of the array.}

\item{i_age}{The index for the age dimension of the array,
or \code{NULL} if the array does not have an age dimension.}
}
\value{
An object of class \code{SpecIterAccount}.
}
\description{
Create an object of class \code{SpecIterAccount} that
contains the information needed to construct
account iterators for an array of population counts.
}
\examples{
x <- SpecIterAccount(dim = c(4, 2, 3),
                     i_time = 3,
                     i_age = 1)
class(x)
}
\seealso{
To create a account iterator from an object of class
\code{SpecIterAccount}, use function
\code{\link[=account]{iter_create_account}}.
}
