% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpecIter-generators.R
\name{SpecIterCollapse}
\alias{SpecIterCollapse}
\title{Create an object of class "SpecIterCollapse"}
\usage{
SpecIterCollapse(dim_self, dim_oth, map_dim, map_pos)
}
\arguments{
\item{dim_self}{An integer vector. The dimensions of array \code{self}.}

\item{dim_oth}{An integer vector. The dimensions of array \code{oth}.}

\item{map_dim}{An integer vector mapping the dimensions of \code{self}
on to those of \code{oth}. If a dimension from \code{self} is not
found in \code{oth}, the associated element of \code{map_dim} is \code{0}.}

\item{map_pos}{A list of integers vectors. Each integer vector
maps positions along a dimension of \code{self} on to a position
along the matching dimension of \code{oth}.}
}
\value{
An object of class \code{SpecIterCollapse}.
}
\description{
Create an object of class \code{SpecIterCollpse} that
contains the information needed to create an iterator that
maps cells in array \code{self} into cells in array \code{oth}.
}
\examples{
x <- SpecIterCollapse(dim_self = c(4L, 2L, 3L),
                      dim_oth = c(4L, 3L),
                      map_dim = c(1L, 0L, 2L),
                      map_pos = list(1:4, c(0L, 0L), 1:3))
class(x)
}
\seealso{
To create a collapse iterator from an object of class
\code{SpecIterCollapse}, use function
\code{\link[=collapse]{create_iter_collapse}}.
}
