% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpecIter-generators.R
\name{SpecIterIncrement}
\alias{SpecIterIncrement}
\title{Create an object of class "SpecIterIncrement"}
\usage{
SpecIterIncrement(
  dim_self,
  dim_oth,
  map_dim,
  comp_type_self,
  i_triangle_self = NULL,
  indices_orig_self = NULL,
  indices_dest_self = NULL,
  i_direction_self = NULL
)
}
\arguments{
\item{dim_self}{An integer vector. The dimensions of array \code{self}.}

\item{dim_oth}{An integer vector. The dimensions of array \code{oth}.}

\item{map_dim}{An integer vector mapping the dimensions of \code{self}
on to those of \code{oth}. If a dimension from \code{self} is not
found in \code{oth}, the associated element of \code{map_dim} is \code{0}.}

\item{comp_type_self}{The type of component: \code{"increment"},
\code{"decrement"}, \code{"orig-dest"}, or \code{"pool"}.}

\item{i_triangle_self}{The index of any dimension of array \code{self}
that has dimtype \code{"triangle"}. Equal to \code{NULL} if \code{self}
does not have a triangle dimension.}

\item{indices_orig_self}{An integer vector containing
the indices of any dimensions of \code{self} with dimtype
\code{"origin"}.}

\item{indices_dest_self}{An integer vector containing
the indices of any dimensions of \code{self} with dimtype
\code{"destination"}.}

\item{i_direction_self}{The index of any dimension of array \code{self}
that has dimtype \code{"direction"}. Equal to \code{NULL} if \code{self}
does not have a direction dimension.}
}
\value{
An object of class \code{SpecIterIncrement}.
}
\description{
Create an object of class \code{SpecIterCollpse} that
contains the information needed to create an iterator
that maps cells in array \code{self} to cells
in arrays of increments or decrements.
}
\examples{
x <- SpecIterIncrement(dim_self = c(4, 2, 3),
                       dim_oth = c(4, 2, 3),
                       map_dim = c(1, 2, 3),
                       comp_type_self = "increment",
                       i_triangle_self = 2)
class(x)
}
\seealso{
To create an increment iterator from an object of class
\code{SpecIterIncrement}, use function
\code{\link[=increment]{create_iter_increment}}.
}
